Form Submission Email Start
Added Email structure for the admin and the client. Currently all the styles are in a separate JS component this is because styles need to be inline for nodemailer as far as I am aware. This was my current solution, I will look more into it. Both emails send and have styling. Error handling needs to be added and the form component could be cleaned up a bit but this is a start.