Form Submission Fix
There was an error when a user tabbed over the "select" inputs, the form would technically still be invalid. This is confusing because the form shows a default value, but that value isn't actually passed. My current fix was to set the validation for these select input to true always since the user has to select an option and then address the inevitable empty string with a conditional statement in the POST request. Not the best solution and I might change it later, but for now this works.