Navigation and UI
Redesigned the navigation for responsive screen size and modified the button appearance while also disabling the form submission button until the form is valid. Changed the navigation hover to fade the other links out, but might want to change it so that the whole footer or header fades out. Also added a social media section for the footer.